requirements.txt
Django==2.2.2
psycopg2-binary==2.8.2
pycrypto==2.6.1
pytz==2019.1
sqlparse==0.3.0
